UpdateBreakingChange
Chuck Lantz
There was a crazy breaking change made to the 14 image.
https://github.com/docker-library/openjdk/issues/429
We have a pre-built image now as well so this wouldn't happen going forward -- very good to know.
Grab the devcontainer.json and Dockerfile from here: https://github.com/microsoft/vscode-dev-containers/tree/master/containers/java/.devcontainer
You won't need the rest. Update the argument in devcontainer.json to select JDK 14. It defaults to 11.
One note on JDK 14 as well is that it's not a LTS version, so generally 11 is a safer starting point unless you need it.